move: guards를 src 하위로 옮김
