+++ src/drivers/driver.module.ts
@@ -1,8 +1,8 @@
-import { GuardModule } from '#auth/guard.module.js';
 import { DriverController } from '#drivers/driver.controller.js';
 import { DriverRepository } from '#drivers/driver.repository.js';
 import { DriverService } from '#drivers/driver.service.js';
 import { DBModule } from '#global/db.module.js';
+import { GuardModule } from '#guards/guard.module.js';
 import { Module } from '@nestjs/common';
 
 @Module({
