+++ src/global/app.module.ts
@@ -1,5 +1,4 @@
 import { AuthModule } from '#auth/auth.module.js';
-import { GuardModule } from '#auth/guard.module.js';
 import jwtConfig from '#configs/jwt.config.js';
 import { postgresConfig } from '#configs/postgres.config.js';
 import { DriverModule } from '#drivers/driver.module.js';
@@ -10,6 +9,7 @@ import { DevController } from '#global/dev.controller.js';
 import { LogInterceptor } from '#global/interceptors/log.interceptor.js';
 import { AlsMiddleware } from '#global/middlewares/als.middleware.js';
 import { StorageModule } from '#global/storage.module.js';
+import { GuardModule } from '#guards/guard.module.js';
 import { MoveModule } from '#move/move.module.js';
 import { QuestionModule } from '#questions/question.module.js';
 import { RequestModule } from '#requests/request.module.js';
