+++ src/guards/access-token.guard.ts
@@ -1,15 +1,16 @@
-import { AuthInvalidRefreshTokenException, AuthInvalidTokenException } from '#auth/auth.exception.js';
-import { GuardService } from '#auth/guard.service.js';
-import { InternalServerErrorException } from '#exceptions/http.exception.js';
+import { AuthInvalidAccessTokenException, AuthInvalidTokenException } from '#auth/auth.exception.js';
+import { InternalServerErrorException, UnauthorizedException } from '#exceptions/http.exception.js';
+import { GuardService } from '#guards/guard.service.js';
 import { IStorage } from '#types/common.types.js';
 import loggingError from '#utils/loggingError.js';
-import { CanActivate, ExecutionContext, Injectable, UnauthorizedException } from '@nestjs/common';
+import { CanActivate, ExecutionContext, Injectable } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
 import { JsonWebTokenError, JwtService } from '@nestjs/jwt';
 import { AsyncLocalStorage } from 'async_hooks';
+import { Request } from 'express';
 
 @Injectable()
-export class RefreshTokenGuard implements CanActivate {
+export class AccessTokenGuard implements CanActivate {
   constructor(
     private readonly jwtService: JwtService,
     private readonly guardService: GuardService,
@@ -21,26 +22,22 @@ export class RefreshTokenGuard implements CanActivate {
     const jwtSecret = this.configService.get('jwtSecret');
 
     const request = context.switchToHttp().getRequest();
-    // NOTE 중복 쿠키 거르는 작업
-    // 쿠키가 이중으로 잡히는 경우가 있어서 사용함
-    const cookies = request.headers.cookie.split(';');
-    const tokens = cookies.filter(cookie => cookie.trim().startsWith('refreshToken=')).map(cookie => cookie.trim().split('=')[1]);
-    const token = tokens[tokens.length - 1];
+    const token = this.extractTokenFromHeader(request);
     if (!token) throw new UnauthorizedException();
 
     try {
       const payload = await this.jwtService.verifyAsync(token, { secret: jwtSecret });
       if (!payload.userId) {
-        throw new AuthInvalidRefreshTokenException();
+        throw new AuthInvalidAccessTokenException();
       }
       const user = await this.guardService.validateUser(payload.userId);
       if (!user) {
-        throw new AuthInvalidRefreshTokenException();
+        throw new AuthInvalidAccessTokenException();
       }
 
       const storage = this.als.getStore();
       Object.assign(storage, payload);
-      storage.refreshToken = token;
+      storage.accessToken = token;
       storage.user = user;
     } catch (err) {
       loggingError(err);
@@ -52,5 +49,10 @@ export class RefreshTokenGuard implements CanActivate {
     }
 
     return true;
+  }
+
+  private extractTokenFromHeader(request: Request) {
+    const [type, token] = request.headers.authorization?.split(' ') ?? [];
+    return type === 'Bearer' ? token : undefined;
   }
 }
