+++ src/guards/guard.module.ts
@@ -1,10 +1,10 @@
-import { GuardService } from '#auth/guard.service.js';
+import { AccessTokenGuard } from '#guards/access-token.guard.js';
+import { GuardService } from '#guards/guard.service.js';
+import { HashPasswordGuard } from '#guards/hash-password.guard.js';
+import { RefreshTokenGuard } from '#guards/refresh-token.guard.js';
 import { UserModule } from '#users/user.module.js';
 import { Module } from '@nestjs/common';
 import { JwtService } from '@nestjs/jwt';
-import { AccessTokenGuard } from './guards/access-token.guard.js';
-import { HashPasswordGuard } from './guards/hash-password.guard.js';
-import { RefreshTokenGuard } from './guards/refresh-token.guard.js';
 
 const guards = [AccessTokenGuard, RefreshTokenGuard, HashPasswordGuard];
