+++ src/guards/refresh-token.guard.ts
@@ -1,16 +1,15 @@
-import { AuthInvalidAccessTokenException, AuthInvalidTokenException } from '#auth/auth.exception.js';
-import { GuardService } from '#auth/guard.service.js';
-import { InternalServerErrorException, UnauthorizedException } from '#exceptions/http.exception.js';
+import { AuthInvalidRefreshTokenException, AuthInvalidTokenException } from '#auth/auth.exception.js';
+import { InternalServerErrorException } from '#exceptions/http.exception.js';
+import { GuardService } from '#guards/guard.service.js';
 import { IStorage } from '#types/common.types.js';
 import loggingError from '#utils/loggingError.js';
-import { CanActivate, ExecutionContext, Injectable } from '@nestjs/common';
+import { CanActivate, ExecutionContext, Injectable, UnauthorizedException } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
 import { JsonWebTokenError, JwtService } from '@nestjs/jwt';
 import { AsyncLocalStorage } from 'async_hooks';
-import { Request } from 'express';
 
 @Injectable()
-export class AccessTokenGuard implements CanActivate {
+export class RefreshTokenGuard implements CanActivate {
   constructor(
     private readonly jwtService: JwtService,
     private readonly guardService: GuardService,
@@ -22,22 +21,26 @@ export class AccessTokenGuard implements CanActivate {
     const jwtSecret = this.configService.get('jwtSecret');
 
     const request = context.switchToHttp().getRequest();
-    const token = this.extractTokenFromHeader(request);
+    // NOTE 중복 쿠키 거르는 작업
+    // 쿠키가 이중으로 잡히는 경우가 있어서 사용함
+    const cookies = request.headers.cookie.split(';');
+    const tokens = cookies.filter(cookie => cookie.trim().startsWith('refreshToken=')).map(cookie => cookie.trim().split('=')[1]);
+    const token = tokens[tokens.length - 1];
     if (!token) throw new UnauthorizedException();
 
     try {
       const payload = await this.jwtService.verifyAsync(token, { secret: jwtSecret });
       if (!payload.userId) {
-        throw new AuthInvalidAccessTokenException();
+        throw new AuthInvalidRefreshTokenException();
       }
       const user = await this.guardService.validateUser(payload.userId);
       if (!user) {
-        throw new AuthInvalidAccessTokenException();
+        throw new AuthInvalidRefreshTokenException();
       }
 
       const storage = this.als.getStore();
       Object.assign(storage, payload);
-      storage.accessToken = token;
+      storage.refreshToken = token;
       storage.user = user;
     } catch (err) {
       loggingError(err);
@@ -49,10 +52,5 @@ export class AccessTokenGuard implements CanActivate {
     }
 
     return true;
-  }
-
-  private extractTokenFromHeader(request: Request) {
-    const [type, token] = request.headers.authorization?.split(' ') ?? [];
-    return type === 'Bearer' ? token : undefined;
   }
 }
