+++ src/move/move.module.ts
@@ -1,5 +1,5 @@
-import { GuardModule } from '#auth/guard.module.js';
 import { DBModule } from '#global/db.module.js';
+import { GuardModule } from '#guards/guard.module.js';
 import { MoveController } from '#move/move.controller.js';
 import { MoveRepository } from '#move/move.repository.js';
 import { MoveService } from '#move/move.service.js';
